requirements.txt
tensorflow==2.4.1
lxml==4.6.3
Pillow==8.2.0
